Address comments: move submesh creation to generator.py
